+++ src/components/bottom-bar/bottom-bar.model.ts
@@ -1,0 +1,4 @@
+export enum ButtonPosition {
+  right = 'right',
+  left = 'left',
+}